Parallel access to the same file is not working, so I made it serial
No matter what I tried, the parallel access to the same file for writing different variables never actually worked.

I think the only way around that issue is to go ahead and have an initialization and a finalization for these routines, which until now I've been trying to avoid. But it's either that or write one file per process. I'll try this again with an initialize and finalize and see if that actually works.